minor fix GTK NIP

+++ database/migrations/2023_02_01_045926_create_gtk_table.php
@@ -16,7 +16,7 @@ class CreateGTKTable extends Migration
         Schema::create('gtk', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->integer('nip')->unique()->nullable();
+            $table->bigInteger('nip')->unique()->nullable();
             $table->string('tempat_lahir');
             $table->date('tanggal_lahir');
             $table->enum('jenis_kelamin', ['laki_laki', 'perempuan']);
